Improve render time by converting complex selectors into reducers.
Generating streams from other data is a relatively
complex transformation. Previously this was being
performed in a selector which placed it in the
render cycle and slowed performance significantly.
This was most evident on the performance of redux-forms.

With the stream generation moved into a reducer the
generation of streams can be controlled and only
carried out when needed.